Create README.md
# TestPurpose

Test purpose is an app that helps other people get from one
location to the other using API's routing and additional methods to help.

HTML: I designed the structure of the html to take on an interesting form, so 
that it was an easy opprtunity for one person t log in adn get what they need.
the structure has a border, Google Maps and a credit card slot along with information systems built inside to decipher 
what method and place the person wants to go. Along with what mode of transporation they want to use.

Javascript: has been implemented inside of the html to make buttons like bike, car, penguin and walk, prompt 
something that works properly.
On top of this, the javascript is linked to the deivce to show that the routing, when node is being use get what
the user wants, which is latitude and longitude.

SQL: Database was implemented to shwo the people that have logged in and that the database is running efficiently.

Node package is running.
Express package is running.
Body-Parser is running
A few other packages are running as well.